+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-This is the first alpha release of SmolChat.
+This release:
+
+- adds a button to stop response generation
+- disables chat options when no chat is selected